+++ src/components/new-tab-link/index.js
@@ -1,0 +1,3 @@
+import main from './new-tab-link';
+
+export default main;